Add fake raw attributes for links
